v1.9
Summary of changes:
- Implemented Turret Tag Lock feature with AprilTag tracking using HuskyLens.
- Optimized codebase performance:
    - Conditional color sensing in StorageSubsystem to reduce I2C overhead.
    - Lazy initialization for HuskyLens in OuttakeSubsystem.
    - Cached follower properties (isBusy, getPose) in all main loops to minimize redundant calls.
- Reorganized telemetry into logical sections across all OpModes (TeleOp and Auto).
- Cleaned up legacy code and improved maintainability.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/config/subsystem/IntakeSubsytem.java
@@ -16,7 +16,7 @@ public class IntakeSubsytem {
         intakeMotor.setPower(0);
     }
 
-    public void setPower(float power) {
+    public void setPower(double power) {
         intakeMotor.setPower(power);
     }
